requirements.txt
Flask>=3.0.0
cryptography>=42.0.0
pytest>=7.0.0
# post-quantum via liboqs — requires native liboqs shared library
# macOS: brew install cmake liboqs (then rebuild liboqs with -DBUILD_SHARED_LIBS=ON)
# Linux: apt install cmake gcc && build from https://github.com/open-quantum-safe/liboqs
liboqs-python>=0.14.1
